refactor: replace model_promote.py with promote_model.py to streamline model promotion process in W&B
